Mock.delay to control how quickly requests finish
A corrolary to Fog::Mock.delay

Default value of 1 means request will finish on the 1st poll.

0 could mean the request finishes before even polling, but that's not
implemented yet (and not necessarily needed either)

+++ lib/ey-core/requests/get_request.rb
@@ -18,27 +18,35 @@ class Ey::Core::Client
       request = self.find(:requests, id)
       request["callback_url"] = "#{self.url}/requests/#{id}/callback"
 
-      response_hash = if request["finished_at"].nil? # finish a previously unfinished request
-                        resource_key, key, resource, resource_url = request.delete("resource")
+      request["poll_count"] ||= 0
+      request["poll_count"] += 1
 
-                        if resource # provisioning
-                          if resource.is_a?(Array)
-                            resource.each { |res| self.data[resource_key][res["id"]] = res }
+      response_hash = if request["finished_at"].nil?
+                        if request["poll_count"] >= delay
+                          # finish a previously unfinished request
+                          resource_key, key, resource, resource_url = request.delete("resource")
+
+                          if resource # provisioning
+                            if resource.is_a?(Array)
+                              resource.each { |res| self.data[resource_key][res["id"]] = res }
+                            else
+                              resource_url ||= if resource.respond_to?(:call)
+                                                 resource.call(request)
+                                               else
+                                                 self.data[resource_key][key] = resource
+                                                 resource.delete("resource_url")
+                                               end
+                            end
+                            request["finished_at"] = Time.now
+                            request["resource"] = resource_url
                           else
-                            resource_url ||= if resource.respond_to?(:call)
-                                               resource.call(request)
-                                             else
-                                               self.data[resource_key][key] = resource
-                                               resource.delete("resource_url")
-                                             end
+                            resource = self.data[resource_key].delete(key) # deprovisioning
+                            self.data[:deleted][resource_key][key] ||= resource
                           end
-                          request["finished_at"] = Time.now
-                          request["resource"] = resource_url
+                          request
                         else
-                          resource = self.data[resource_key].delete(key) # deprovisioning
-                          self.data[:deleted][resource_key][key] ||= resource
+                          request
                         end
-                        request
                       elsif request["finished_at"] # already finished
                         request
                       else
@@ -50,6 +58,9 @@ class Ey::Core::Client
                           end
                         end
                       end
+
+      response_hash = response_hash.dup
+      response_hash.delete("poll_count")
 
       response(
         :body   => {"request" => response_hash},
